Optimise uploads time
